feat(owner): rename /admin/payout → /owner, redesign as car owner economics dashboard (#88)
- Guard /api/admin/summary with requireAdminOrOwner (security fix)
- Create /owner page accessible to any car owner (not admin-only)
- Redirect /admin/payout → /owner
- Wire /owner into nav for admin/owner users
- Extract BreakEvenCard and RateAssistant to components/ for reuse
- Add beMetrics unit tests (9 cases covering the fleet economics math)
