+++ app/admin/layout.tsx
@@ -8,7 +8,7 @@ import { useT } from "@/components/locale-provider";
 import { useReservations, useOwnerCarShorts } from "./_shared";
 import { useMe } from "@/hooks/use-me";
 
-const OWNER_PAGES = ["/admin", "/admin/hygiene", "/admin/settlement", "/admin/payout"];
+const OWNER_PAGES = ["/admin", "/admin/hygiene", "/admin/settlement", "/owner"];
 
 function SubNav() {
   const t = useT();
@@ -31,7 +31,7 @@ function SubNav() {
     { href: "/admin/members", label: t("admin.sub_members") },
     { href: "/admin/hygiene", label: t("admin.sub_data") },
     { href: "/admin/settlement", label: t("admin.sub_settlement") },
-    { href: "/admin/payout", label: t("admin.sub_payout") },
+    { href: "/owner", label: t("owner.title") },
     { href: "/admin/settings", label: t("admin.sub_settings") },
   ];
